+++ CoreDB/dbo/Tables/SourceType.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[SourceType]
 (
-	[Id] INT NOT NULL,
+	[Id] BIGINT NOT NULL,
 	[Name] nvarchar(500) NOT NULL,
 	CONSTRAINT [PK_SourceType] PRIMARY KEY CLUSTERED ([Id] ASC)
 )
